+++ treehugger/cli/__init__.py
@@ -1,14 +1,13 @@
 # -*- coding:utf-8 -*-
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-import argparse
-import os
 import sys
 
 from six.moves import shlex_quote
 
 from .decrypt_file import decrypt_file
 from .edit import edit
+from .execute import execute
 from .encrypt_file import encrypt_file
 from .parser import parser, subparsers
 from .. import yaml
@@ -28,41 +27,8 @@ command_funcs = {
     'decrypt-file': decrypt_file,
     'encrypt-file': encrypt_file,
     'edit': edit,
+    'exec': execute,
 }
-
-
-exec_parser = subparsers.add_parser(
-    'exec',
-    description='''
-        Execute a command with environment variables set to the decryption of
-        the variables stored in Treehugger YAML in EC2 User Data, or a file.
-    ''',
-)
-exec_parser.add_argument('-f', '--file', type=str, default=None, dest='filename',
-                         help='The path to the file to use for environment variables, as opposed to EC2 User Data')
-exec_parser.add_argument('command', nargs=argparse.REMAINDER)
-
-
-def execute(args):
-    command = args.command
-    if command and command[0] == '--':
-        command = command[1:]
-    if not command:
-        print('No command to execute provided', file=sys.stderr)
-        raise SystemExit(1)
-
-    if args.filename:
-        data = yaml.load_file_or_die(args.filename)
-        env_dict = EnvironmentDict.from_yaml_dict(data)
-    else:
-        data = load_user_data_as_yaml_or_die()
-        env_dict = EnvironmentDict.from_yaml_dict(data)
-    unencrypted_env_dict = env_dict.decrypt_all_encrypted(plain=True)
-    os.environ.update(unencrypted_env_dict)
-    os.execlp(command[0], *command)
-
-
-command_funcs['exec'] = execute
 
 
 print_parser = subparsers.add_parser(
